third_party/blink/web_tests/fast/forms/required-attribute-001-expected.txt

There are three form control elements below, all required.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS requiredFor("input") is true
PASS requiredFor("textarea") is true
PASS requiredFor("select") is true
PASS successfullyParsed is true

TEST COMPLETE

  
X
